Update README.md
a languege model can do simple talking task it is the first transformer base model i build When I was 16, I called her luna . The code is a basic implementation of a language model algorithm. You can use your dataset to train it directly. However, it is not recommended because this is not the most efficient model.